% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_data.R
\name{copy_data}
\alias{copy_data}
\title{Copy tutorial data to the Desktop}
\usage{
copy_data()
}
\value{
None
}
\description{
Copies the data used in the R for Immunologists tutorial to the Desktop
}
\examples{
r4immunologists::copy_data()

}
